+++ tests/unit/composables/events/ReloadSecretsEvent.spec.js
@@ -1,0 +1,9 @@
+import ReloadSecretsEvent from 'src/composables/events/ReloadSecretsEvent';
+import { Subject } from 'rxjs';
+
+describe('Test composable: ReloadSecretsEvent', () => {
+  it('should export a Subject', () => {
+    expect(ReloadSecretsEvent).toBeDefined();
+    expect(ReloadSecretsEvent).toEqual(new Subject());
+  });
+});