+++ changelog.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## [Unreleased]
+
+### Added
+
+* Setup AI settings page with:
+  * Table to display all secrets.
+  * Add action to create a secret.
+  * Add action to edit a secret.
+  * Add action to remove a secret. 
+
 ## [1.0.0] - 2024/10/15
 
 ### Added